I found a way. THE BEST WAY  :: Click here to read -->
After a lot of trial an error, I found a way to transfer players IMMEDIATLEY (besides network lagg). This method is the first I've seen. I won't explain how it works. The code is hashed and don't even
 think about trying to steal it. When you execute the transferEvent implemented, the MCPE client will have no delay. There will be no lagg. You just BAM hit another server! Hopefully I can implement this soon. For now, it is a single file. WARNING: This method is not stable yet. Do not use this. If you do, just setup like this:
```
<?php
use pocketmine\...
...

use LostTransfer\InstaTransfer as IT;

...code...
/*
Then,
*/
$lostTransferEvent->transferInstant($player, lr.game-server.cc, 19132);
```
Thank you for reading all of this geez. this took me forever to write. xD